fix rating dialog that does not dismiss
