+++ presentation/viewModel/src/main/kotlin/com/berlin/aflami/viewmodel/details/common/CommonUiState.kt
@@ -1,6 +1,7 @@
 package com.berlin.aflami.viewmodel.details.common
 
 import androidx.compose.runtime.Immutable
+import com.berlin.aflami.viewmodel.base.ErrorUiState
 import com.berlin.viewModel.R
 
 @Immutable
@@ -21,6 +22,21 @@ data class CompanyProductionUiState(
     val name: String = "",
     val country: String = "",
 )
+
+@Immutable
+data class SnackBarUiState(
+    val isVisible: Boolean = false,
+    val isOperationSucceeded: Boolean = false,
+    val errorUiState: ErrorUiState = ErrorUiState(),
+    val snackBarStatus: SNACK_BAR_STATUS? = null,
+)
+
+enum class SNACK_BAR_STATUS {
+    ADD_MOVIE_TO_LIST,
+    CREATE_NEW_LIST,
+    LIST_DELETED,
+    LIST_RENAMED,
+}
 
 fun <T> Set<T>.toggle(item: T): Set<T> =
     if (contains(item)) this - item else this + item
