+++ presentation/viewModel/src/main/kotlin/com/berlin/aflami/viewmodel/details/series/TVShowDetailsUiState.kt
@@ -1,5 +1,6 @@
 package com.berlin.aflami.viewmodel.details.series
 
+import com.berlin.aflami.viewmodel.details.common.SnackBarUiState
 import com.berlin.aflami.viewmodel.shareduistate.ActorUiState
 import com.berlin.aflami.viewmodel.shareduistate.TVShowUiState
 
@@ -7,7 +8,7 @@ data class TVShowDetailsUiState(
     val isFavorite: Boolean = false,
     val posters: List<String> = emptyList(),
     val isTVShowHasVideo: Boolean = false,
-    val videoUrl: String?= null ,
+    val videoUrl: String?= null,
     val tvShowUiState: TVShowUiState = TVShowUiState(),
     val isDescriptionExpanded: Boolean = false,
     val castList: List<ActorUiState> = emptyList(),
@@ -25,6 +26,7 @@ data class TVShowDetailsUiState(
     val snackBarMessage: String? = null,
     val isSnackBarStatusSuccess: Boolean? = null,
     val isNotSupportedFeatureDialogVisible: Boolean = false,
+    val snackBar: SnackBarUiState = SnackBarUiState(),
 )
 
 data class TVShowDetailsTabsUiState(
