+++ presentation/viewModel/src/main/kotlin/com/berlin/aflami/viewmodel/list/ListScreenState.kt
@@ -3,7 +3,7 @@ package com.berlin.aflami.viewmodel.list
 import androidx.compose.runtime.Immutable
 import androidx.compose.ui.text.input.TextFieldValue
 import androidx.paging.PagingData
-import com.berlin.aflami.viewmodel.details.movie.SnackBarUiState
+import com.berlin.aflami.viewmodel.details.common.SnackBarUiState
 import com.berlin.aflami.viewmodel.reusableinteractionlistener.list.addTiList.FavouriteListItemUiState
 import com.berlin.aflami.viewmodel.reusableinteractionlistener.list.createnewlist.CreateNewListUiState
 import com.berlin.aflami.viewmodel.reusableinteractionlistener.list.editList.EditListSheetState
